requirements.txt
streamlit==1.29.0
openai==1.6.1
pandas==2.1.4
PyPDF2==3.0.1
python-dotenv==1.0.0
openpyxl==3.1.2

